LPS-92088 Move search integration tests from portal-impl to portal-search-test module
