+++ modules/apps/portal-search/portal-search-test/src/testIntegration/java/com/liferay/portal/search/indexer/test/IndexerRegistryUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search;
+package com.liferay.portal.search.indexer.test;
 
 import com.liferay.portal.kernel.model.Contact;
 import com.liferay.portal.kernel.model.Organization;
