+++ modules/apps/portal-search/portal-search-test/src/testIntegration/java/com/liferay/portal/search/test/BackupAndRestoreIndexesTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search;
+package com.liferay.portal.search.test;
 
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.search.IndexAdminHelperUtil;
