+++ modules/apps/portal-search/portal-search-test/src/testIntegration/java/com/liferay/portal/search/test/ElasticsearchRegistrationTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch;
+package com.liferay.portal.search.test;
 
 import com.liferay.portal.kernel.search.SearchEngine;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
